requirements.txt
torch
torchvision
numpy
pandas
opencv-python
matplotlib
Pillow
streamlit
